backend/requirements.txt
# NeuraMed v5.0 Backend

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
torch>=2.1.0
torchvision>=0.16.0
efficientnet-pytorch==0.7.1
Pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
